Get ECampus Colors by Name

This function retrieves one or more color hex codes from the ecampus_palettes dataset.

Usage

get_ecampus_colors(...)

Arguments

...

One or more color names as character strings.

Value

A named vector of hex codes for the requested colors.

Examples

get_ecampus_colors(c("eCore_Green", "eCore_Teal"))
#> eCore_Green  eCore_Teal 
#>   "#7BA640"   "#00A9A6" 
get_ecampus_colors("eCore_OER_Orange", palette = "USG-eCore")
#> Warning: Color 'USG-eCore' not found in ecampus_palettes.
#> eCore_OER_Orange        USG-eCore 
#>        "#F6931E"               "" 
get_ecampus_colors("eMajor_Navy", palette = "all") # Search in all branches
#> Warning: Color 'all' not found in ecampus_palettes.
#> eMajor_Navy         all 
#>   "#0D4860"          ""